test parameterization

+++ tests/frontend/test_engine.py
@@ -44,7 +44,7 @@ batch_engines = [
     sf.Engine("gaussian", backend_options={"batch_size": 2, "cutoff_dim": 6}),
     sf.Engine("fock", backend_options={"batch_size": 2, "cutoff_dim": 6}),
     sf.Engine("tf", backend_options={"batch_size": 2, "cutoff_dim": 6}),
-    sf.Engine("tf", backend_options={"batch_size": 2, "cutoff_dim": 6, "dtype": tf.complex64}),
+    sf.Engine("tf", backend_options={"batch_size": 2, "cutoff_dim": 6, "dtype": tf.complex128}),
 ]
 
 engines = [
